+++ src/adb/command/host-transport/exec.ts
@@ -1,0 +1,14 @@
+import Command from '../../command';
+import { Duplex } from 'stream';
+import WithToString from '../../../WithToString';
+
+export default class ExecCommand extends Command<Duplex> {
+  async execute(command: string | ArrayLike<WithToString>): Promise<Duplex> {
+    if (Array.isArray(command)) {
+      command = command.map(this.escape).join(' ');
+    }
+    this._send(`exec:${command}`);
+    await this.readOKAY();
+    return this.parser.raw();
+  }
+}